Switch examples to have separate crates for API
