+++ examples/future_api/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "future_api"
+version = "0.1.0"
+authors = ["Adam Wright <shaladdle@gmail.com>"]
+
+[dependencies]
