+++ examples/future_cli/src/main.rs
@@ -1,7 +1,7 @@
 extern crate clap;
-extern crate tarpc;
-extern crate future_starter;
 extern crate futures;
+extern crate hello_api;
+extern crate tarpc;
 extern crate tokio_core;
 
 use futures::Future;
@@ -16,7 +16,7 @@ fn main() {
     let addr = matches.value_of("server_address").unwrap().first_socket_addr();
     let person_name = matches.value_of("person_name").unwrap();
     let mut reactor = tokio_core::reactor::Core::new().unwrap();
-    reactor.run(future_starter::FutureClient::connect(addr, Options::default())
+    reactor.run(hello_api::FutureClient::connect(addr, Options::default())
             .map_err(tarpc::Error::from)
             .and_then(|client| client.hello(person_name.into()))
             .map(|resp| println!("{}", resp)))
