+++ examples/hello_api/src/lib.rs
@@ -1,0 +1,9 @@
+#![feature(plugin)]
+#![plugin(tarpc_plugins)]
+
+#[macro_use]
+extern crate tarpc;
+
+service! {
+    rpc hello(name: String) -> String;
+}
