+++ examples/sync_cli/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "sync_cli"
+version = "0.1.0"
+authors = ["Adam Wright <shaladdle@gmail.com>"]
+
+[dependencies]
+clap = "2.2"
+hello_api = { path = "../hello_api" }
+tarpc = { path = "../../" }
